Implemented signal queue, now signals are stored in a queue structure.

+++ Aurora/include/string.h
@@ -30,10 +30,9 @@ AU_EXTERN AU_EXPORT char* strchr (char* str, int character);
 //! M E M O R Y  G L O B A L   F U N C T I O N S
 //!=====================================================
 AU_EXTERN AU_EXPORT void memset(void *targ, uint8_t val, uint32_t len);
-AU_EXTERN AU_EXPORT void  memcpy(void *targ, void *src, uint32_t len);
+AU_EXTERN AU_EXPORT void  memcpy(void *targ, void *src, size_t len);
 AU_EXTERN AU_EXPORT  int memcmp(const void *first, const void *second, size_t length);
 AU_EXTERN AU_EXPORT void* memmove (void*, const void*, size_t);
-
 
 
 #endif
